feat(AIAI-575): improve history JSON parsing and UI entity view models

+++ smart-workflow-dashboard/src/com/axonivy/utils/smart/workflow/governance/ui/entity/CaseHistoryGroup.java
@@ -48,6 +48,6 @@ public class CaseHistoryGroup implements HistoryGroupView {
 
   @Override public String getLastUpdatedText() { return stats.getLastUpdatedText(); }
   @Override public int    getMessageCount()    { return stats.getMessageCount(); }
-  @Override public int    getTotalTokens()     { return stats.getTotalTokens(); }
+  @Override public long   getTotalTokens()     { return stats.getTotalTokens(); }
   @Override public String getModelName()       { return stats.getModelName(); }
 }
